Support multi-material approaches in Nalu
Use cases for multiphysics requires multi-matrial. This is mostly
due to conjugate heat transfer applications that will use one single
realm.

* Simple restructuring of Material parsing:

******** Newly supported multi-material approach:

    material_properties:

      properties:

        - name: mat_one
          target_name: [block_10]

          specifications:

            - name: density
              type: constant
              value: 1.0

            - name: specific_heat
              type: constant
              value: 1.0

            - name: thermal_conductivity
              type: constant
              value: 1.0

        - name: mat_two
          target_name: [block_20]

          specifications:

            - name: density
              type: constant
              value: 1.0

            - name: specific_heat
              type: constant
              value: 1.0

            - name: thermal_conductivity
              type: constant
              value: 1.0

******** Routine block is still supported (this work is concentrating on a temperature-form approach)

    material_properties:

      target_name: [block_10, block_20]
      specifications:
        - name: density
          type: constant
          value: 1.0

        - name: specific_heat
          type: constant
          value: 1.0

        - name: thermal_conductivity
          type: constant
          value: 1.0

* Transition parsing to the new design. For old material_block, create a single
material property object.

* Support backward compatibility. Things get odd when attempting multi-material
enthalpy - something that is not a use case of interest (immediately, at least).
Protect by a throw should enthalpy system have multiple material evaluators found.

* Make sure that the material property algorithms extract the proper promoted
higher-order part.

* add two block use case to DG/HC

Notes:

a) No diffs as expected.
b) May consider moving the material evaluator from nodes to the element..
