+++ include/MaterialPropertys.h
@@ -9,13 +9,12 @@
 #ifndef MaterialPropertys_h
 #define MaterialPropertys_h
 
-#include <Enums.h>
+#include "Enums.h"
 
 // yaml for parsing..
 #include <yaml-cpp/yaml.h>
 
-#include <map>
-#include <string>
+// basic c++
 #include <vector>
 
 namespace YAML {
@@ -27,9 +26,6 @@ namespace nalu{
 
 class Realm;
 class MaterialProperty;
-class MaterialPropertyData;
-class ReferencePropertyData;
-class PropertyEvaluator;
 class Simulation;
 
 typedef std::vector<MaterialProperty *> MaterialPropertyVector;
@@ -52,16 +48,9 @@ public:
   Realm *parent();  
 
   Realm &realm_;
-  MaterialPropertyVector materialPropertyVector_;
-  std::string propertyTableName_;
 
-  // vectors and maps required to manage full set of options
+  MaterialPropertyVector materialPropertyVector_;
   std::vector<std::string> targetNames_;
-  std::map<std::string, double> universalConstantMap_;
-  std::map<PropertyIdentifier, MaterialPropertyData*> propertyDataMap_;
-  std::map<std::string, ReferencePropertyData*> referencePropertyDataMap_; /* defines overall species ordering */
-  std::map<PropertyIdentifier, PropertyEvaluator*> propertyEvalMap_;
-  std::map<std::string, ReferencePropertyData*> tablePropertyMap_;
 };
