+++ src/Realm.C
@@ -33,6 +33,7 @@
 #include "FieldTypeDef.h"
 #include "LinearSystem.h"
 #include "master_element/MasterElement.h"
+#include "MaterialProperty.h"
 #include "MaterialPropertys.h"
 #include "MeshMotionInfo.h"
 #include "NaluParsing.h"
@@ -1134,45 +1135,50 @@ Realm::setup_initial_conditions()
 void
 Realm::setup_property()
 {
-  // loop over all target names
-  const std::vector<std::string> targetNames = get_physics_target_names();
-  for (size_t itarget=0; itarget < targetNames.size(); ++itarget) {
-
-    // target need not be subsetted since nothing below will depend on topo
-    stk::mesh::Part *targetPart = metaData_->get_part(targetNames[itarget]);
-
-    // loop over propertyMap
-    std::map<PropertyIdentifier, ScalarFieldType *>::iterator ii;
-    for ( ii=propertyMap_.begin(); ii!=propertyMap_.end(); ++ii ) {
-
-      // extract property id and field pointer
-      PropertyIdentifier thePropId = (*ii).first;
-      ScalarFieldType *thePropField = (*ii).second;
-
-      // find the material property data object
-      MaterialPropertyData *matData = NULL;
-      std::map<PropertyIdentifier, MaterialPropertyData*>::iterator itf =
-        materialPropertys_.propertyDataMap_.find(thePropId);
-      if ( itf == materialPropertys_.propertyDataMap_.end() ) {
-        // will need to throw
-        NaluEnv::self().naluOutputP0() << "issue with property: " << PropertyIdentifierNames[thePropId] << std::endl;
-        throw std::runtime_error("Please add property specification ");
-      }
-      else {
-        matData = (*itf).second;
-      }
-
-      switch( matData->type_) {
+  // loop overall material property blocks
+  for ( size_t i = 0; i < materialPropertys_.materialPropertyVector_.size(); ++i ) {
+    
+    MaterialProperty *matPropBlock = materialPropertys_.materialPropertyVector_[i];
+    
+    // loop over all target names
+    for (size_t itarget=0; itarget < matPropBlock->targetNames_.size(); ++itarget) {
+      
+      // target need not be subsetted since nothing below will depend on topo
+      const std::string physicsPartName = physics_part_name(matPropBlock->targetNames_[itarget]);
+      stk::mesh::Part *targetPart = metaData_->get_part(physicsPartName);
+      
+      // loop over propertyMap
+      std::map<PropertyIdentifier, ScalarFieldType *>::iterator ii;
+      for ( ii=propertyMap_.begin(); ii!=propertyMap_.end(); ++ii ) {
+        
+        // extract property id and field pointer
+        PropertyIdentifier thePropId = (*ii).first;
+        ScalarFieldType *thePropField = (*ii).second;
+        
+        // find the material property data object
+        MaterialPropertyData *matData = NULL;
+        std::map<PropertyIdentifier, MaterialPropertyData*>::iterator itf =
+          matPropBlock->propertyDataMap_.find(thePropId);
+        if ( itf == matPropBlock->propertyDataMap_.end() ) {
+          // will need to throw
+          NaluEnv::self().naluOutputP0() << "issue with property: " << PropertyIdentifierNames[thePropId] << std::endl;
+          throw std::runtime_error("Please add property specification ");
+        }
+        else {
+          matData = (*itf).second;
+        }
+        
+        switch( matData->type_) {
 
         case CONSTANT_MAT:
         {
-
+          
           // for constant specific heat, proceed in specialty code; create the appropriate enthalpy evaluator
           if (thePropId == SPEC_HEAT_ID && needsEnthalpy_) {
             // extract reference temperature
             double tRef = 300.0;
-            extract_universal_constant("reference_temperature", tRef, true);
-
+            matPropBlock->extract_universal_constant("reference_temperature", tRef, true);
+            
             // set up evaluators required for all cases
             PropertyEvaluator *theCpPropEval = NULL;
             PropertyEvaluator *theEnthPropEval = NULL;
@@ -1218,8 +1224,8 @@ Realm::setup_property()
 
             }
             // push to prop eval
-            materialPropertys_.propertyEvalMap_[SPEC_HEAT_ID] = theCpPropEval;
-            materialPropertys_.propertyEvalMap_[ENTHALPY_ID]  = theEnthPropEval;
+            matPropBlock->propertyEvalMap_[SPEC_HEAT_ID] = theCpPropEval;
+            matPropBlock->propertyEvalMap_[ENTHALPY_ID]  = theEnthPropEval;
           }
           else {
 
@@ -1280,7 +1286,7 @@ Realm::setup_property()
 
                 // all props will use Tref
                 double tRef = 0.0;
-                extract_universal_constant("reference_temperature", tRef, true);
+                matPropBlock->extract_universal_constant("reference_temperature", tRef, true);
 
                 if ( uniformFlow_ ) {
                   // props computed based on YkRef and Tref
@@ -1301,7 +1307,7 @@ Realm::setup_property()
                 if ( uniformFlow_ ) {
                   // props computed based on YkRef and T
                   viscPropEval = new SutherlandsPropertyEvaluator(
-                    materialPropertys_.referencePropertyDataMap_, matData->polynomialCoeffsMap_ );
+                    matPropBlock->referencePropertyDataMap_, matData->polynomialCoeffsMap_ );
                 }
                 else {
                   // props computed based on Yk and T
@@ -1315,7 +1321,7 @@ Realm::setup_property()
               }
 
               // create the property alg and push to evalmap
-              materialPropertys_.propertyEvalMap_[thePropId] = viscPropEval;
+              matPropBlock->propertyEvalMap_[thePropId] = viscPropEval;
 
             }
             break;
@@ -1330,7 +1336,7 @@ Realm::setup_property()
             {
               // R
               double universalR = 8314.4621;
-              extract_universal_constant("universal_gas_constant", universalR, true);
+              matPropBlock->extract_universal_constant("universal_gas_constant", universalR, true);
 
               // create the property alg and push to evalmap
               PropertyEvaluator *theCpPropEval = NULL;
@@ -1338,19 +1344,19 @@ Realm::setup_property()
               if ( uniformFlow_ ) {
                 // props computed based on reference values
                 theCpPropEval = new SpecificHeatPropertyEvaluator(
-                    materialPropertys_.referencePropertyDataMap_, matData->lowPolynomialCoeffsMap_,
+                    matPropBlock->referencePropertyDataMap_, matData->lowPolynomialCoeffsMap_,
                     matData->highPolynomialCoeffsMap_, universalR);
                 theEnthPropEval = new EnthalpyPropertyEvaluator(
-                    materialPropertys_.referencePropertyDataMap_, matData->lowPolynomialCoeffsMap_,
+                    matPropBlock->referencePropertyDataMap_, matData->lowPolynomialCoeffsMap_,
                     matData->highPolynomialCoeffsMap_, universalR);
               }
               else {
                 // props computed based on transported Yk values
                 theCpPropEval = new SpecificHeatTYkPropertyEvaluator(
-                    materialPropertys_.referencePropertyDataMap_, matData->lowPolynomialCoeffsMap_,
+                    matPropBlock->referencePropertyDataMap_, matData->lowPolynomialCoeffsMap_,
                     matData->highPolynomialCoeffsMap_, universalR, *metaData_);
                 theEnthPropEval = new EnthalpyTYkPropertyEvaluator(
-                   materialPropertys_.referencePropertyDataMap_, matData->lowPolynomialCoeffsMap_,
+                   matPropBlock->referencePropertyDataMap_, matData->lowPolynomialCoeffsMap_,
                    matData->highPolynomialCoeffsMap_, universalR, *metaData_);
               }
 
@@ -1360,8 +1366,8 @@ Realm::setup_property()
               propertyAlg_.push_back(auxAlg);
 
               // set property maps...
-              materialPropertys_.propertyEvalMap_[SPEC_HEAT_ID] = theCpPropEval;
-              materialPropertys_.propertyEvalMap_[ENTHALPY_ID] = theEnthPropEval;
+              matPropBlock->propertyEvalMap_[SPEC_HEAT_ID] = theCpPropEval;
+              matPropBlock->propertyEvalMap_[ENTHALPY_ID] = theEnthPropEval;
             }
             break;
 
@@ -1377,7 +1383,7 @@ Realm::setup_property()
         
             // everyone will require R
             double universalR = 8314.4621;
-            extract_universal_constant("universal_gas_constant", universalR, true);
+            matPropBlock->extract_universal_constant("universal_gas_constant", universalR, true);
             
             // create the property evaluator
             PropertyEvaluator *rhoPropEval = NULL;
@@ -1385,8 +1391,8 @@ Realm::setup_property()
               // load mw and reference species
               std::vector<std::pair<double, double> > mwMassFracVec;
               std::map<std::string, ReferencePropertyData*>::const_iterator itrp;
-              for ( itrp = materialPropertys_.referencePropertyDataMap_.begin();
-                    itrp!= materialPropertys_.referencePropertyDataMap_.end(); ++itrp) {
+              for ( itrp = matPropBlock->referencePropertyDataMap_.begin();
+                    itrp!= matPropBlock->referencePropertyDataMap_.end(); ++itrp) {
                 ReferencePropertyData *propData = (*itrp).second;
                 std::pair<double,double> thePair;
                 thePair = std::make_pair(propData->mw_,propData->massFraction_);
@@ -1394,7 +1400,7 @@ Realm::setup_property()
               }
               if ( IDEAL_GAS_T_MAT == matData->type_ ) {
                 double pRef = 101325.0;
-                extract_universal_constant("reference_pressure", pRef, true);
+                matPropBlock->extract_universal_constant("reference_pressure", pRef, true);
                 rhoPropEval = new IdealGasTPropertyEvaluator(pRef, universalR, mwMassFracVec);
               }
               else {
@@ -1405,14 +1411,14 @@ Realm::setup_property()
               // load mw
               std::vector<double> mwVec;
               std::map<std::string, ReferencePropertyData*>::const_iterator itrp;
-              for ( itrp = materialPropertys_.referencePropertyDataMap_.begin();
-                    itrp!= materialPropertys_.referencePropertyDataMap_.end(); ++itrp) {
+              for ( itrp = matPropBlock->referencePropertyDataMap_.begin();
+                    itrp!= matPropBlock->referencePropertyDataMap_.end(); ++itrp) {
                 ReferencePropertyData *propData = (*itrp).second;
                 mwVec.push_back(propData->mw_);
               }
               if ( IDEAL_GAS_T_MAT == matData->type_ ) {
                 double pRef = 101325.0;
-                extract_universal_constant("reference_pressure", pRef, true);
+                matPropBlock->extract_universal_constant("reference_pressure", pRef, true);
                 rhoPropEval = new IdealGasTYkPropertyEvaluator(pRef, universalR, mwVec, *metaData_);
               }
               else {
@@ -1421,7 +1427,7 @@ Realm::setup_property()
             }
 
             // push back property evaluator to map
-            materialPropertys_.propertyEvalMap_[thePropId] = rhoPropEval;
+            matPropBlock->propertyEvalMap_[thePropId] = rhoPropEval;
 
             // create the property algorithm
             TemperaturePropAlgorithm *auxAlg
@@ -1443,15 +1449,15 @@ Realm::setup_property()
             double pRef = 101325.0;
             double tRef = 300.0;
             double universalR = 8314.4621;
-            extract_universal_constant("reference_pressure", pRef, true);
-            extract_universal_constant("reference_temperature", tRef, true);
-            extract_universal_constant("universal_gas_constant", universalR, true);
+            matPropBlock->extract_universal_constant("reference_pressure", pRef, true);
+            matPropBlock->extract_universal_constant("reference_temperature", tRef, true);
+            matPropBlock->extract_universal_constant("universal_gas_constant", universalR, true);
 
             // load mw
             std::vector<double> mwVec;
             std::map<std::string, ReferencePropertyData*>::const_iterator itrp;
-            for ( itrp = materialPropertys_.referencePropertyDataMap_.begin();
-                  itrp!= materialPropertys_.referencePropertyDataMap_.end(); ++itrp) {
+            for ( itrp = matPropBlock->referencePropertyDataMap_.begin();
+                  itrp!= matPropBlock->referencePropertyDataMap_.end(); ++itrp) {
               ReferencePropertyData *propData = (*itrp).second;
               mwVec.push_back(propData->mw_);
             }
@@ -1460,7 +1466,7 @@ Realm::setup_property()
             PropertyEvaluator *rhoPropEval = new IdealGasYkPropertyEvaluator(pRef, tRef, universalR, mwVec, *metaData_);
 
             // push back property evaluator to map
-            materialPropertys_.propertyEvalMap_[thePropId] = rhoPropEval;
+            matPropBlock->propertyEvalMap_[thePropId] = rhoPropEval;
 
             // create the property algorithm
             GenericPropAlgorithm *auxAlg
@@ -1486,7 +1492,7 @@ Realm::setup_property()
         case HDF5_TABLE_MAT:
         {
 	  if ( HDF5ptr_ == NULL ) {
-	    HDF5ptr_ = new HDF5FilePtr( materialPropertys_.propertyTableName_ );
+	    HDF5ptr_ = new HDF5FilePtr( matPropBlock->propertyTableName_ );
 	  }
 
  	  // create the new TablePropAlgorithm that knows how to read from HDF5 file
@@ -1527,73 +1533,51 @@ Realm::setup_property()
 	break;
 
       case GENERIC: 
-        { 
-          // default property evaluator
-          PropertyEvaluator *propEval = NULL;
-          
-          // extract the property evaluator name
-          std::string propEvalName = matData->genericPropertyEvaluatorName_;
-
-          if ( propEvalName == "water_viscosity_T" ) {
-            propEval = new WaterViscosityTPropertyEvaluator(*metaData_);
-          }
-          else if ( propEvalName == "water_density_T" ) {
-            propEval = new WaterDensityTPropertyEvaluator(*metaData_);
-          }
-          else if ( propEvalName == "water_specific_heat_T" ) {
-            propEval = new WaterSpecHeatTPropertyEvaluator(*metaData_);
-            // create the enthalpy prop evaluator and store
-            WaterEnthalpyTPropertyEvaluator *theEnthPropEval 
-              = new WaterEnthalpyTPropertyEvaluator(*metaData_);
-            materialPropertys_.propertyEvalMap_[ENTHALPY_ID] = theEnthPropEval;
-          }
-          else if ( propEvalName == "water_thermal_conductivity_T" ) {
-            propEval = new WaterThermalCondTPropertyEvaluator(*metaData_);
-          }
-          else {
-            throw std::runtime_error("Realm::setup_property: unknown GENERIC type: " + propEvalName);
-          }
-          
-          // for now, all of the above are TempPropAlgs; push it back
-          TemperaturePropAlgorithm *auxAlg
-            = new TemperaturePropAlgorithm( *this, targetPart, thePropField, propEval);
-          propertyAlg_.push_back(auxAlg);
-
-          // push back property evaluator to map
-          materialPropertys_.propertyEvalMap_[thePropId] = propEval;
+      { 
+        // default property evaluator
+        PropertyEvaluator *propEval = NULL;
+        
+        // extract the property evaluator name
+        std::string propEvalName = matData->genericPropertyEvaluatorName_;
+        
+        if ( propEvalName == "water_viscosity_T" ) {
+          propEval = new WaterViscosityTPropertyEvaluator(*metaData_);
         }
-        break;
-
+        else if ( propEvalName == "water_density_T" ) {
+          propEval = new WaterDensityTPropertyEvaluator(*metaData_);
+        }
+        else if ( propEvalName == "water_specific_heat_T" ) {
+          propEval = new WaterSpecHeatTPropertyEvaluator(*metaData_);
+          // create the enthalpy prop evaluator and store
+          WaterEnthalpyTPropertyEvaluator *theEnthPropEval 
+            = new WaterEnthalpyTPropertyEvaluator(*metaData_);
+          matPropBlock->propertyEvalMap_[ENTHALPY_ID] = theEnthPropEval;
+        }
+        else if ( propEvalName == "water_thermal_conductivity_T" ) {
+          propEval = new WaterThermalCondTPropertyEvaluator(*metaData_);
+        }
+        else {
+          throw std::runtime_error("Realm::setup_property: unknown GENERIC type: " + propEvalName);
+        }
+        
+        // for now, all of the above are TempPropAlgs; push it back
+        TemperaturePropAlgorithm *auxAlg
+          = new TemperaturePropAlgorithm( *this, targetPart, thePropField, propEval);
+        propertyAlg_.push_back(auxAlg);
+        
+        // push back property evaluator to map
+        matPropBlock->propertyEvalMap_[thePropId] = propEval;
+      }
+      break;
+      
         case MaterialPropertyType_END:
           break;
-
+          
         default:
           throw std::runtime_error("Realm::setup_property: unknown type:");
+        }
       }
     }
-  }
-}
-
-//--------------------------------------------------------------------------
-//-------- extract_universal_constant --------------------------------------
-//--------------------------------------------------------------------------
-void
-Realm::extract_universal_constant( 
-  const std::string name, double &value, const bool useDefault)
-{
-  std::map<std::string, double >::iterator it
-    = materialPropertys_.universalConstantMap_.find(name);
-  if ( it == materialPropertys_.universalConstantMap_.end() ) {
-    // not found
-    if ( useDefault ) {
-      NaluEnv::self().naluOutputP0() << "WARNING: Reference value for " << name << " not found " << " using " << value << std::endl;
-    }
-    else {
-      throw std::runtime_error("Realm::setup_property: reference value not found: " + name);
-    }
-  }
-  else {
-    value = (*it).second;
   }
 }
 
@@ -4390,17 +4374,20 @@ Realm::get_nc_alg_current_normal()
 //--------------------------------------------------------------------------
 //-------- get_material_prop_eval ------------------------------------------
 //--------------------------------------------------------------------------
-PropertyEvaluator *
+void
 Realm::get_material_prop_eval(
-  const PropertyIdentifier thePropID)
+  const PropertyIdentifier thePropID,
+  std::vector<PropertyEvaluator*> &propEvalVec)
 {
-  PropertyEvaluator *thePropEval = NULL;
-  std::map<PropertyIdentifier, PropertyEvaluator*>::const_iterator iter
-    = materialPropertys_.propertyEvalMap_.find(thePropID);
-  if (iter != materialPropertys_.propertyEvalMap_.end()) {
-    thePropEval = (*iter).second;
+  for ( size_t i = 0; i < materialPropertys_.materialPropertyVector_.size(); ++i ) {
+    PropertyEvaluator *thePropEval = NULL;
+    std::map<PropertyIdentifier, PropertyEvaluator*>::const_iterator iter
+      = materialPropertys_.materialPropertyVector_[i]->propertyEvalMap_.find(thePropID);
+    if (iter != materialPropertys_.materialPropertyVector_[i]->propertyEvalMap_.end()) {
+      thePropEval = (*iter).second;
+    }
+    propEvalVec.push_back(thePropEval);
   }
-  return thePropEval;
 }
 
 //--------------------------------------------------------------------------
